Add Feature for Show Edit
