+++ src/components/EditModal/EditModal.module.scss
@@ -1,0 +1,6 @@
+@import '../../styles/break-points.scss';
+
+.Modal {
+   @include Modal;
+
+}
