+++ src/index.module.scss
@@ -8,9 +8,12 @@ body {
 
   height: 100vh;
   width: 100vw;
+  
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+  monospace;
 }
+
+
