+++ src/styles/break-points.scss
@@ -3,6 +3,29 @@ $laptop-size: 769px;
 $desktop-size: 1025px;
 $desktop-large-size: 1201px;
 
+@mixin Modal {
+    position : absolute;
+    display : flex;
+    justify-content: center;
+    align-items: center;
+    overflow-y: auto;
+
+    width : 100%;
+    height : 100%;
+    top : 0;
+    left : 0;
+
+    background-color: rgb(37, 37, 37);
+
+    img {
+      width : 50px;
+      position : absolute;
+
+      top : 15px;
+      left : 5vw;
+    }
+}
+
 @mixin tablet {
   @media screen and (min-width: #{$tablet-size}) {
     @content;
