+++ src/types/Modal.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react'
+
+export interface ModalProps {
+    status : Boolean;
+    setStatus : Dispatch<SetStateAction<Boolean>>;
+}
